INPROGRESS - Combat
Combat now starts and ends. Refactored combat class so that there is just one class and no longer an interface, this has created some hacking with the companion character and enemies also ignore player.
Player is considered invincible and lacked certain scripts, due to time factor and also purpose of the demo I feel this okay at this stage though could be changed in the future.